requirements.txt
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
kafka-python==2.0.2
redis==5.0.1
websockets==12.0
requests